feat: создать необходимую логику в модуле gateway

+++ gateway/src/main/java/ru/practicum/shareit/booking/dto/BookItemRequestDto.java
@@ -12,7 +12,7 @@ import java.time.LocalDateTime;
 @NoArgsConstructor
 @AllArgsConstructor
 public class BookItemRequestDto {
-	private long itemId;
+	private int itemId;
 	@FutureOrPresent
 	private LocalDateTime start;
 	@Future
